feat[inclusion tree improvement]: Add BaseException.h file containing basic exceptions, to remove threading layer dependency upon xpcf injection headers. Change Fifo dependency to  BaseException.h
